adds show routes dynamic with student id, need to make show data dynamic to each specific student
