+++ src/Header.jsx
@@ -2,7 +2,7 @@ export function Header() {
   return (
     <header>
       <nav>
-        <a href="#">Home</a> | <a href="#">Link</a> | <a href="#">Login to Student Portal</a>
+        <a href="/">Home</a> | <a href="#">Link</a> | <a href="#">Login to Student Portal</a>
       </nav>
     </header>
   );
