+++ src/StudentsShow.jsx
@@ -1,7 +1,9 @@
 export function StudentsShow(props) {
   return (
     <div>
-      <h1> Students Resume </h1>
+      <h1> Student Resume </h1>
+      {props.students[0].first_name}
+      <p>{props.students[0].id}</p>
     </div>
   );
 }
